Lottery: improved UI and options flow

+++ database/migrations/2025_11_19_162133_create_unit_preferences_table.php
@@ -13,7 +13,7 @@ return new class () extends Migration {
             $table->id();
             $table->foreignIdFor(Family::class)->constrained()->cascadeOnDelete();
             $table->foreignIdFor(Unit::class)->constrained()->cascadeOnDelete();
-            $table->unsignedTinyInteger('order'); // 1-based ordering
+            $table->unsignedInteger('order'); // 1-based ordering
             $table->timestamps();
 
             $table->unique(['family_id', 'unit_id']);
